+++ src/main/java/racingcar/dao/RacingCarDao.java
@@ -17,7 +17,6 @@ public class RacingCarDao {
     private final JdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert insertActor;
 
-    @Autowired
     public RacingCarDao(final JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
         this.insertActor = new SimpleJdbcInsert(jdbcTemplate)
