+++ src/main/resources/data.sql
@@ -1,6 +1,6 @@
 -- TODO: 기능 구현에 필요한 내용을 추가하거나 수정하세요.
 
-CREATE TABLE PLAY_RESULT
+CREATE TABLE IF NOT EXISTS PLAY_RESULT
 (
     id          INT UNSIGNED NOT NULL AUTO_INCREMENT,
     winners     VARCHAR(50) NOT NULL,
@@ -9,7 +9,7 @@ CREATE TABLE PLAY_RESULT
     PRIMARY KEY (id)
 );
 
-CREATE TABLE LOG
+CREATE TABLE IF NOT EXISTS LOG
 (
     id      INT         NOT NULL AUTO_INCREMENT,
     game_id INT         NOT NULL,
